+++ pyodide-build/pyodide_build/buildpkg.py
@@ -136,6 +136,7 @@ def get_bash_runner():
             "NUMPY_LIB",
             "PYODIDE_PACKAGE_ABI",
             "HOSTINSTALLDIR",
+            "TARGETINSTALLDIR",
             "HOSTSITEPACKAGES",
             "PYMAJOR",
             "PYMINOR",
@@ -145,6 +146,7 @@ def get_bash_runner():
             "SIDE_MODULE_LDFLAGS",
             "STDLIB_MODULE_CFLAGS",
             "OPEN_SSL_ROOT",
+            "UNISOLATED_PACKAGES",
         ]
     } | {"PYODIDE": "1"}
     if "PYODIDE_JOBS" in os.environ:
@@ -454,6 +456,7 @@ def package_wheel(
     srcpath: Path,
     build_metadata: dict[str, Any],
     bash_runner: BashRunnerWithSharedEnvironment,
+    host_install_dir,
 ):
     """Package a wheel
 
@@ -507,6 +510,18 @@ def package_wheel(
         if result.returncode != 0:
             print("ERROR: post failed")
             exit_with_stdio(result)
+
+    python_dir = f"python{sys.version_info.major}.{sys.version_info.minor}"
+    host_site_packages = Path(host_install_dir) / f"lib/{python_dir}/site-packages"
+    if build_metadata.get("cross-build-env"):
+        subprocess.check_call(
+            ["pip", "install", "-t", str(host_site_packages), f"{name}=={ver}"]
+        )
+
+    cross_build_files: list[str] | None = build_metadata.get("cross-build-files")
+    if cross_build_files:
+        for file in cross_build_files:
+            shutil.copy((wheel_dir / file), host_site_packages / file)
 
     test_dir = distdir / "tests"
     nmoved = 0
@@ -753,11 +768,7 @@ def build_package(
             )
         if not sharedlibrary:
             package_wheel(
-                name,
-                pkg_root,
-                srcpath,
-                build_metadata,
-                bash_runner,
+                name, pkg_root, srcpath, build_metadata, bash_runner, host_install_dir
             )
 
         shutil.rmtree(pkg_root / "dist", ignore_errors=True)