+++ docs/project/changelog.md
@@ -56,6 +56,8 @@ substitutions:
   translated to negative Python ints.
   {pr}`2484`
 
+- {{ Fix }} Fix garbage collection of `once_callable` {pr}`2401`
+
 - {{ Enhancement }} `run_in_pyodide` now has support for pytest assertion
   rewriting and various other improvements.
   {pr}`2510`
@@ -86,6 +88,10 @@ substitutions:
 
 - {{ Fix }} micropip now correctly compares packages with prerelease version
   {pr}`2532`
+
+- {{ Enhancement }} {func}`micropip.install` now accepts a `pre` parameter.
+  If set to `True`, micropip will include pre-release and development versions.
+  {pr}`2542`
 
 ### Packages